requirements.txt
streamlit
scikit-learn
pandas
numpy==1.23.4
requests
Pillow==9.2.0
torch==1.13.0
torchvision==0.14.0
